app/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.6
django-bootstrap4==2.3.1
django-colorfield==0.4.1
django-copyright==1.0.0
django-crispy-forms==1.11.0
django-geojson==3.1.0
django-grappelli==2.14.3
django-leaflet==0.27.1
django-registration==3.1.1
django-tables2==2.3.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2021.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
django-filter==2.4.0
certifi==2020.12.5
chardet==3.0.4
idna==2.9
requests==2.23.0
urllib3==1.25.8
django-redis==4.12.1
django-recaptcha==2.0.6
social-auth-app-django==4.0.0